replace animals worktag from fishing with manualskilled
It seem to be related to what work is allowed (based e.g. on backstory),
and animals worktag seems to be taking care of animals, which fishing is not.
Manualskilled worktag seems to be about non-trivial manual work,
and since even things like plant cutting have the tag, it seems to fit here
too.